Fix the interface include directory on the logger target (#3)
The install interface directory should just be `include/`, not the path
to where the file is installed. In the default case the former will be
one level deeper `include/<project_name>`, and in the general case will
be `include/${LOGGER_HEADER_DIR}` which is not what we want since
consumers will always specify include paths relative to the root e.g.
`#include <${project_name}/logger.hpp>` or similar.