Fix spdlog handling (#7)
For the moment we must continue imposing the requirement of finding
spdlog on consumers of any package that uses rapids-logger, even those
that only link to it privately, in order to safely support header-only
consumers of rapids-logger like rmm and raft that must pass the
requirement onto their consumers for the impl target to be compiled. We
should try to relax this eventually so that core requirements like rmm
don't wind up placing this burden transitively onto everything that
consumes them, but in order to do that we will have to make a
significant breaking build change by forcing all consumers to find
spdlog themselves, so that change is better made as a near-final step
once we have sorted out a few more of the outstanding issues around
rapids-logger. This change requires moving the find call into the
`rapids_make_logger` function so that the logger is added to every
export set that creates it rather than just the first one to load the
module. This PR also fixes a typo in the export sets that were being
used for rapids-logger.